Wrote script to clean up LaTeX build files
Used Shellcheck to change all my scripts to bash and lint them. Exported my
scripts to $PATH. Fixed my liked Arch packages.

+++ setup.sh
@@ -1,6 +1,6 @@
-#!/bin/zsh
+#!/bin/bash
 
-setup_marble=true
+# setup_marble=true
 
 # vars {{{ #
 
@@ -8,7 +8,6 @@ linux_config_dir=/home/kevin/linux-config
 dot_dir=$linux_config_dir/dotfiles
 config_dir=$linux_config_dir/configs
 sys_config_dir=~/.config
-pc=$(hostname)
 
 # }}} vars #
 
@@ -19,8 +18,8 @@ pc=$(hostname)
 # Get directory variables from script.
 # List of dotfiles I want to link to system.
 dotfiles=(bashrc gitconfig inputrc imwheelrc vimrc xinitrc Xmodmap Xresources zshrc)
-cd $dot_dir
-for dotfile in $dotfiles; do
+cd $dot_dir || exit
+for dotfile in "${dotfiles[@]}"; do
 	# Remove system dotfile.
     rm ~/."$dotfile"
 	# Link dotfile in repository to system dotfile.
@@ -43,30 +42,29 @@ ln -s "$dot_dir"/nvim "$orig_file" && echo "Setup Neovim config."
 
 # configs {{{ #
 
-cd $linux_config_dir
+cd $linux_config_dir || exit
 # List of config dirs in system config dir that I want to link to.
 config_dirs=(neofetch i3 i3blocks sam-i3blocks ranger)
-# config_dirs=(neofetch i3 i3blocks ranger)
-for cur_dir in $config_dirs; do
+for cur_dir in "${config_dirs[@]}"; do
 	# Remove system file.
-	rm -rf "$sys_config_dir"/"$cur_dir"
+	rm -rf ${sys_config_dir:?}/"$cur_dir"
 	# Link config file in repository to system config location.
 	ln -s "$config_dir"/"$cur_dir" "$sys_config_dir"/"$cur_dir" && echo "Linked $cur_dir config."
 done
 
 # }}} configs #
 
-#  marble mouse {{{ # 
+# #  marble mouse {{{ #
 
-if [[ "$setup_marble" = true ]]; then
-	config_file_name="10-evdev.conf"
-	sys_layout_file="/etc/X11/xorg.conf.d/$config_file_name"
-	config_dir="$linux_config_dir/x/marble-mouse"
+# if [[ "$setup_marble" = true ]]; then
+	# config_file_name="10-evdev.conf"
+	# sys_layout_file="/etc/X11/xorg.conf.d/$config_file_name"
+	# config_dir="$linux_config_dir/x/marble-mouse"
 
-	# Remove system file.
-	sudo rm -rf "$sys_layout_file"
-	# Link config file in repository to system config location.
-	sudo ln -s "$config_dir"/"$config_file_name" "$sys_layout_file" && echo "Setup Marble mouse config."
-fi
+	# # Remove system file.
+	# sudo rm -rf "$sys_layout_file"
+	# # Link config file in repository to system config location.
+	# sudo ln -s "$config_dir"/"$config_file_name" "$sys_layout_file" && echo "Setup Marble mouse config."
+# fi
 
-#  }}} marble mouse # 
+# #  }}} marble mouse #
